Unificación de .gitignore y limpieza de archivos innecesarios y configuracion de carpetas rutas y endpoins de inicio de sesion
