+++ TaskFlow_API/TaskFlow/urls/auth_urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+from ..views.auth_views import CustomTokenObtainPairView
+from rest_framework_simplejwt.views import TokenRefreshView
+
+
+urlpatterns = [
+    path('api/token/', CustomTokenObtainPairView.as_view(), name='token_obtain_pair'),
+    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
+]
