--- TaskFlow_API/TaskFlow/views.py
@@ -1,13 +1,0 @@
-from django.shortcuts import render
-from rest_framework import viewsets
-from .serializer import UserSerializer,CustomTokenObtainPairSerializer
-from django.contrib.auth.models import User
-from rest_framework_simplejwt.views import TokenObtainPairView
-# Create your views here.
-
-class UserView(viewsets.ModelViewSet):
-    serializer_class = UserSerializer
-    queryset = User.objects.all()
-
-class CustomTokenObtainPairView(TokenObtainPairView):
-    serializer_class = CustomTokenObtainPairSerializer
