+++ TaskFlow_API/TaskFlow/views/auth_views.py
@@ -1,0 +1,6 @@
+from ..serializer import CustomTokenObtainPairSerializer
+from rest_framework_simplejwt.views import TokenObtainPairView
+# Create your views here.
+
+class CustomTokenObtainPairView(TokenObtainPairView):
+    serializer_class = CustomTokenObtainPairSerializer
