+++ TaskFlow_API/TaskFlow_API/urls.py
@@ -19,5 +19,5 @@ from django.urls import path,include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('TaskFlow/',include('TaskFlow.urls'))
+    path('auth/', include('TaskFlow.urls.auth_urls')),
 ]
